Move some parameter utils to `metabase-lib` (#25863)
* Move parameter target utils to metabase-lib

* Move parameter-type utils to metabase-lib

* Move parameter constants to metabase-lib

* Use Parameter interface, not UiParameter type, in operator utils

* Move operator utils to metabase-lib

* Move filter utils to metabase-lib
