+++ frontend/src/metabase-lib/lib/parameters/utils/operators.ts
@@ -1,12 +1,15 @@
-import { UiParameter } from "metabase/parameters/types";
+import { Parameter } from "metabase-types/types/Parameter";
 
 import {
   doesOperatorExist,
   getOperatorByTypeAndName,
 } from "metabase-lib/lib/operators/utils";
 import { NUMBER, STRING, PRIMARY_KEY } from "metabase-lib/lib/types/constants";
-import { PARAMETER_OPERATOR_TYPES } from "../constants";
-import { getParameterType, getParameterSubType } from "./parameter-type";
+import {
+  getParameterType,
+  getParameterSubType,
+} from "metabase-lib/lib/parameters/utils/parameter-type";
+import { PARAMETER_OPERATOR_TYPES } from "metabase-lib/lib/parameters/constants";
 
 type OperatorType = "date" | "number" | "string";
 
@@ -28,7 +31,7 @@ export function getParameterOperatorName(maybeOperatorName?: string) {
   return doesOperatorExist(maybeOperatorName) ? maybeOperatorName : "=";
 }
 
-export function deriveFieldOperatorFromParameter(parameter: UiParameter) {
+export function deriveFieldOperatorFromParameter(parameter: Parameter) {
   const type = getParameterType(parameter);
   const subtype = getParameterSubType(parameter);
   const operatorType = getParameterOperatorType(type);
@@ -71,7 +74,7 @@ export function buildTypedOperatorOptions(
   });
 }
 
-export function getNumberParameterArity(parameter: UiParameter) {
+export function getNumberParameterArity(parameter: Parameter) {
   switch (parameter.type) {
     case "number/=":
     case "number/!=":
@@ -83,7 +86,7 @@ export function getNumberParameterArity(parameter: UiParameter) {
   }
 }
 
-export function getStringParameterArity(parameter: UiParameter) {
+export function getStringParameterArity(parameter: Parameter) {
   switch (parameter.type) {
     case "string/=":
     case "string/!=":
