+++ frontend/src/metabase-lib/lib/queries/utils/card.js
@@ -2,9 +2,9 @@ import _ from "underscore";
 import { updateIn } from "icepick";
 
 import { normalizeParameterValue } from "metabase/parameters/utils/parameter-values";
-import { deriveFieldOperatorFromParameter } from "metabase/parameters/utils/operators";
 
 import Utils from "metabase/lib/utils";
+import { deriveFieldOperatorFromParameter } from "metabase-lib/lib/parameters/utils/operators";
 import * as Q_DEPRECATED from "metabase-lib/lib/queries/utils"; // legacy
 
 export const STRUCTURED_QUERY_TEMPLATE = {
