+++ frontend/src/metabase/components/DateRelativeWidget/DateRelativeWidget.tsx
@@ -3,7 +3,7 @@ import { t } from "ttag";
 import cx from "classnames";
 import _ from "underscore";
 
-import { DATE_MBQL_FILTER_MAPPING } from "metabase/parameters/constants";
+import { DATE_MBQL_FILTER_MAPPING } from "metabase-lib/lib/parameters/constants";
 
 type Shortcut = {
   name: string;
