+++ frontend/src/metabase/parameters/components/ParameterValueWidget.jsx
@@ -8,14 +8,6 @@ import {
   getParameterIconName,
   getParameterWidgetTitle,
 } from "metabase/parameters/utils/ui";
-import {
-  isDateParameter,
-  isNumberParameter,
-} from "metabase/parameters/utils/parameter-type";
-import {
-  getNumberParameterArity,
-  getStringParameterArity,
-} from "metabase/parameters/utils/operators";
 
 import PopoverWithTrigger from "metabase/components/PopoverWithTrigger";
 import Icon from "metabase/components/Icon";
@@ -30,6 +22,14 @@ import WidgetStatusIcon from "metabase/parameters/components/WidgetStatusIcon";
 import FormattedParameterValue from "metabase/parameters/components/FormattedParameterValue";
 import NumberInputWidget from "metabase/parameters/components/widgets/NumberInputWidget";
 import StringInputWidget from "metabase/parameters/components/widgets/StringInputWidget";
+import {
+  getNumberParameterArity,
+  getStringParameterArity,
+} from "metabase-lib/lib/parameters/utils/operators";
+import {
+  isDateParameter,
+  isNumberParameter,
+} from "metabase-lib/lib/parameters/utils/parameter-type";
 
 import ParameterFieldWidget from "./widgets/ParameterFieldWidget/ParameterFieldWidget";
 import S from "./ParameterWidget.css";
