+++ frontend/src/metabase/parameters/components/widgets/ParameterFieldWidget/ParameterFieldWidget.jsx
@@ -5,12 +5,12 @@ import { t } from "ttag";
 import _ from "underscore";
 
 import FieldValuesWidget from "metabase/components/FieldValuesWidget";
-import { deriveFieldOperatorFromParameter } from "metabase/parameters/utils/operators";
 import {
   WidgetRoot,
   Footer,
   UpdateButton,
 } from "metabase/parameters/components/widgets/Widget.styled";
+import { deriveFieldOperatorFromParameter } from "metabase-lib/lib/parameters/utils/operators";
 import {
   getFilterArgumentFormatOptions,
   isEqualsOperator,
