+++ frontend/src/metabase/parameters/utils/dashboard-options.js
@@ -1,6 +1,6 @@
 import { t } from "ttag";
-import { ID_OPTION } from "../constants";
-import { buildTypedOperatorOptions } from "./operators";
+import { ID_OPTION } from "metabase-lib/lib/parameters/constants";
+import { buildTypedOperatorOptions } from "metabase-lib/lib/parameters/utils/operators";
 
 export function getDashboardParameterSections() {
   return [
