+++ frontend/src/metabase/parameters/utils/formatting.ts
@@ -1,9 +1,10 @@
 import { ngettext, msgid } from "ttag";
 
 import { formatValue } from "metabase/lib/formatting";
+import { getParameterType } from "metabase-lib/lib/parameters/utils/parameter-type";
 
 import { UiParameter } from "../types";
-import { getParameterType, isFieldFilterParameter } from "./parameter-type";
+import { isFieldFilterParameter } from "./parameter-type";
 import { formatDateValue } from "./date-formatting";
 
 function inferValueType(parameter: UiParameter) {
