+++ frontend/src/metabase/parameters/utils/parameter-type.ts
@@ -1,42 +1,8 @@
 import _ from "underscore";
-import { FIELD_FILTER_PARAMETER_TYPES } from "metabase/parameters/constants";
-
 import { FieldFilterUiParameter } from "metabase/parameters/types";
 import { Parameter } from "metabase-types/types/Parameter";
-
-export function getParameterType(parameter: Parameter | string) {
-  return typeof parameter === "string"
-    ? splitType(parameter)[0]
-    : parameter.sectionId || splitType(parameter)[0];
-}
-
-export function getParameterSubType(parameter: Parameter) {
-  const [, subtype] = splitType(parameter);
-  return subtype;
-}
-
-function splitType(parameterOrType: Parameter | string) {
-  const parameterType = _.isString(parameterOrType)
-    ? parameterOrType
-    : parameterOrType?.type || "";
-
-  return parameterType.split("/");
-}
-
-export function isDateParameter(parameter: Parameter | string) {
-  const type = getParameterType(parameter);
-  return type === "date";
-}
-
-export function isNumberParameter(parameter: Parameter) {
-  const type = getParameterType(parameter);
-  return type === "number";
-}
-
-export function isStringParameter(parameter: Parameter) {
-  const type = getParameterType(parameter);
-  return type === "string";
-}
+import { FIELD_FILTER_PARAMETER_TYPES } from "metabase-lib/lib/parameters/constants";
+import { getParameterType } from "metabase-lib/lib/parameters/utils/parameter-type";
 
 export function isFieldFilterParameter(
   parameter: Parameter,
