+++ frontend/src/metabase/parameters/utils/template-tag-options.js
@@ -4,9 +4,12 @@ import {
   OPTIONS_WITH_OPERATOR_SUBTYPES,
   PARAMETER_OPERATOR_TYPES,
   ID_OPTION,
-} from "../constants";
-import { getOperatorDisplayName, buildTypedOperatorOptions } from "./operators";
-import { fieldFilterForParameter } from "./filters";
+} from "metabase-lib/lib/parameters/constants";
+import {
+  getOperatorDisplayName,
+  buildTypedOperatorOptions,
+} from "metabase-lib/lib/parameters/utils/operators";
+import { fieldFilterForParameter } from "metabase-lib/lib/parameters/utils/filters";
 
 export function getParameterOptions() {
   return [
